fix(rustup): 1.2.0 (7cb9914fc 2015-05-25)
Adjusted method signatures to deal with the recently added `Constness`
flag.

For now it just defaults to `NotConst` which seems a reasonable default.
However, I must admit I have no idea what I was doing, except to make
builds work with the latest nightly (also with `cargo test`).

Unfortunately, this makes the code incompatible with stable, which could
probably be considered a problem.

This PR might just be a hint and for others to see in case they have
the same problems ... and I wouldn't consider it an actual fix.

As a last note: even though this compiles on a recent nightly, in case
you want to use `serde_macros`, these will not work due to other issues
that are beyond my fixing capabilities.