Update for rust-lang/rust#29850
This will of course break the build on stable, as syntex_syntax will
need to be updated as well. The suite now passes on nightly, however.
Though the pull request upstream which broke this adds support for
attributes in statement and expression contexts, I've opted not to
actually support it here yet, as we presumably cannot handle that on
stable for at least 2 more Rust versions.

Fixes #56